d3d: move host writes where they belong
There was a reason for defering (two passes update the same dyn.buffer
in the same frame, those can be accumulated instead of doing a full
DISCARD write twice) but the logic was broken due to optimizing out the
BindGraphicsPipeline in some cases. So keep it simple for now, the
multi-update in same frame case is discouraged anyway.