vk: resource sharing
Sharable resources, which are at least buffers and textures, can now be
orphaned, meaning the owning QRhi can be destroyed before releasing the
QRhiResource. They will still function in a limited way (cannot be rebuilt
anymore) as long as there is at least one QRhi associated with the
QRhiResourceSharingHost (i.e. as long as the share group is non-empty).